replace ExpiringActivityManager with ExpiringActivityManagerV2
