+++ WireFoundation/Sources/WireUtilitiesPackage/ExpiringActivityRegistration/withExpiringActivity.swift
@@ -20,10 +20,6 @@ private let sharedManager = ExpiringActivityManager(
     performer: ExpiringActivityProcessInfoWrapper()
 )
 
-private let sharedManagerV2 = ExpiringActivityManagerV2(
-    performer: ExpiringActivityProcessInfoWrapper()
-)
-
 // MARK: - Non-throwing
 
 public func withExpiringActivity(
@@ -76,21 +72,4 @@ func withExpiringActivity(
     } onCancel: {
         task.cancel()
     }
-}
-
-// MARK: - Throwing v2
-
-public func withExpiringActivityV2(
-    reason: String,
-    block: @escaping @Sendable () async throws -> Void
-) async throws {
-
-    let task = Task(operation: block)
-    await sharedManagerV2.track(reason: reason, task: task)
-    try await withTaskCancellationHandler {
-        try await task.value
-    } onCancel: {
-        task.cancel()
-    }
-
 }
